pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dc-overview"
version = "1.1.1"
description = "GPU Datacenter Monitoring Suite - Prometheus, Grafana & Exporters"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "CryptoLabs", email = "info@cryptolabs.co.za"}
]
keywords = ["gpu", "datacenter", "monitoring", "prometheus", "grafana", "nvidia", "dcgm", "vast.ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "jinja2>=3.0.0",
    "paramiko>=3.0.0",
    # Web application (Docker deployment)
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "gunicorn>=21.0.0",
    "prometheus-client>=0.19.0",
    "werkzeug>=2.3.0",
    # Security
    "flask-wtf>=1.2.0",
    "flask-limiter>=3.5.0",
]

[project.optional-dependencies]
ipmi = [
    "ipmi-monitor>=1.0.0",
]
full = [
    "ipmi-monitor>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/cryptolabsza/dc-overview"
Documentation = "https://cryptolabs.co.za/dc-monitoring/"
Repository = "https://github.com/cryptolabsza/dc-overview"
Issues = "https://github.com/cryptolabsza/dc-overview/issues"
Discord = "https://discord.gg/7yeHdf5BuC"

[project.scripts]
dc-overview = "dc_overview.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dc_overview = [
    "templates/*.yml",
    "templates/*.yaml",
    "templates/*.json",
    "templates/*.service",
    "templates/*.j2",
    "templates/*.html",
    "templates/*.conf",
    "templates/grafana/**/*.yml",
    "templates/grafana/**/*.yaml",
    "web_templates/*.html",
    "dashboards/*.json",
    "dc_exporter/*.c",
    "dc_exporter/*.py",
    "_build_info.py",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]
